update: セッティングとチューニングデータの動作確認

+++ src/settings/settings.controller.ts
@@ -4,28 +4,41 @@ import {
   Post,
   Body,
   Param,
+  Req,
   Put,
   Delete,
   HttpCode,
   HttpStatus,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { SettingsService } from './settings.service';
+import { MyCarService } from 'src/mycar/mycar.service';
 import { CreateSettingDto } from './dto/create-setting.dto';
 import { Setting } from './schemas/setting.schema';
+import { AuthGuard } from '@nestjs/passport';
 
 @Controller('settings')
 export class SettingsController {
-  constructor(private readonly settingsService: SettingsService) {}
+  constructor(
+    private readonly settingsService: SettingsService,
+    private readonly myCarsService: MyCarService,
+  ) {}
 
   @Post()
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(201)
-  async create(@Body() createSettingDto: CreateSettingDto): Promise<Setting> {
-    return this.settingsService.create(createSettingDto);
-  }
-
-  @Get()
-  async findAll(): Promise<Setting[]> {
-    return this.settingsService.findAll();
+  async create(
+    @Req() req,
+    @Body() createSettingDto: CreateSettingDto,
+  ): Promise<Setting> {
+    const userId = req.user.id;
+    const mycarId = createSettingDto.mycarId;
+    if (await this.myCarsService.isUserRelatedToCar(userId, mycarId)) {
+      return this.settingsService.create(createSettingDto);
+    } else {
+      throw new UnauthorizedException("You don't have access to this car.");
+    }
   }
 
   @Get(':id')
@@ -33,18 +46,38 @@ export class SettingsController {
     return this.settingsService.findOne(id);
   }
 
+  @Get('mycar/:mycarId')
+  async findAllByMyCarId(
+    @Param('mycarId') mycarId: string,
+  ): Promise<Setting[]> {
+    return this.settingsService.findAllByMyCarId(mycarId);
+  }
+
   @Put(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.NO_CONTENT)
   async update(
+    @Req() req,
     @Param('id') id: string,
     @Body() updateSettingDto: CreateSettingDto,
   ): Promise<void> {
-    await this.settingsService.update(id, updateSettingDto);
+    const userId = req.user.id;
+    if (await this.settingsService.isUserRelatedToSetting(userId, id)) {
+      await this.settingsService.update(id, updateSettingDto);
+    } else {
+      throw new UnauthorizedException("You don't have access to this setting.");
+    }
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(HttpStatus.NO_CONTENT)
-  async delete(@Param('id') id: string): Promise<void> {
-    await this.settingsService.delete(id);
+  async delete(@Req() req, @Param('id') id: string): Promise<void> {
+    const userId = req.user.id;
+    if (await this.settingsService.isUserRelatedToSetting(userId, id)) {
+      await this.settingsService.delete(id);
+    } else {
+      throw new UnauthorizedException("You don't have access to this setting.");
+    }
   }
 }
